update FilledButton to have a loading option implemented using thunks
