+++ src/elements/index.ts
@@ -2,7 +2,7 @@ export { default as Typography } from "./Typography";
 export * from "./Typography";
 export { default as Dialog } from "./styled/Dialog";
 export { default as DropdownSelect } from "./DropdownSelect";
-export { default as FilledButton } from "./styled/FilledButton";
+export { default as FilledButton } from "./FilledButton";
 export { default as MultiSelect } from "./MultiSelect";
 export { default as OutlinedButton } from "./OutlinedButton";
 export * from "./DropdownSelect";
@@ -12,7 +12,6 @@ export { default as GradientTypography } from "./styled/GradientTypography";
 export * from "./GradientTextInput";
 export { default as HorizontalLine } from "./styled/HorizontalLine";
 export { default as Link } from "./Link";
-export { default as LoadingButton } from "./LoadingButton";
 export { default as TableSkeleton } from "./skeletons/TableSkeleton";
 export { default as Switch } from "./Switch";
 export { default as TextArea } from "./TextArea";
