--- src/elements/styled/FilledButton.ts
@@ -1,31 +1,0 @@
-import { Button, ButtonProps } from "@mui/material";
-import { styled } from "@mui/material/styles";
-import theme from "theme";
-
-interface FilledButtonProps extends ButtonProps {
-  backgroundColor?: string;
-  fontSize?: string;
-  fontWeight?: number;
-}
-
-const FilledButton = styled(Button)<FilledButtonProps>(
-  ({
-    backgroundColor = theme.gradients.music,
-    fontSize = theme.button.fontSize,
-    fontWeight = theme.button.fontWeight,
-    disabled,
-  }) => ({
-    background: backgroundColor,
-    borderRadius: "7px",
-    fontSize: fontSize,
-    fontWeight: fontWeight,
-    lineHeight: theme.button.lineHeight,
-    color: "white",
-    font: theme.typography.button.font,
-    textTransform: "none",
-    padding: "12px 16px",
-    opacity: disabled ? 0.5 : 1,
-  })
-);
-
-export default FilledButton;
