+++ src/modules/song/slice.ts
@@ -1,5 +1,4 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { uploadSong } from "./thunks";
 import { SongState } from "./types";
 
 const initialState: SongState = {
@@ -14,21 +13,12 @@ const songSlice = createSlice({
     receiveSongs(state, { payload }) {
       state.songs = payload;
     },
-  },
-  extraReducers: (builder) => {
-    builder
-      .addCase(uploadSong.pending, (state) => {
-        state.isLoading = true;
-      })
-      .addCase(uploadSong.fulfilled, (state) => {
-        state.isLoading = false;
-      })
-      .addCase(uploadSong.rejected, (state) => {
-        state.isLoading = false;
-      });
+    setIsLoading(state, { payload }) {
+      state.isLoading = payload;
+    },
   },
 });
 
-export const { receiveSongs } = songSlice.actions;
+export const { receiveSongs, setIsLoading } = songSlice.actions;
 
 export default songSlice.reducer;
